chore(llma): rename back to original directory name
