+++ plugin-server/src/ingestion/ai-costs/process-ai-event.ts
