+++ plugin-server/src/ingestion/ai-costs/providers/embeddings.ts
