+++ plugin-server/src/ingestion/ai-costs/providers/llm-provider-costs.json
