+++ plugin-server/src/ingestion/ai-costs/providers/manual-providers.ts
