+++ plugin-server/src/ingestion/ai-costs/scripts/update-ai-costs.ts
